move logic to component
